fix(ci): extract BenchSupport to its own pure-module file
Requiring bench_helper.rb from bench_support_spec.rb executed
`Pgbus.configure { |c| c.stats_enabled = false }` at load time, which
poisoned the global configuration for the rest of the test suite. Later
specs that exercised the executor's stat-recording path (timezone_handling,
dispatcher cleanup_stats) silently no-op'd because `record_stat` returns
early when `config.stats_enabled` is false.

Move BenchSupport to benchmarks/bench_support.rb with zero side effects.
bench_helper.rb (which still mutates Pgbus.configuration for benchmark
scripts) requires the new file; the spec requires only the pure module.
